requirements.txt
streamlit 
pillow 
pytesseract 
requests
google-generativeai
opencv-python
python-dotenv

